move ProfileContainerViewModel into Binders package
