+++ apple/OmnivoreKit/Sources/Views/Utils/MacOSSizeClass.swift
@@ -8,20 +8,20 @@ import SwiftUI
 
 // https://stackoverflow.com/questions/63526478/swiftui-userinterfacesizeclass-for-universal-macos-ios-views
 #if os(macOS)
-  enum UserInterfaceSizeClass {
+  public enum UserInterfaceSizeClass {
     case compact
     case regular
   }
 
-  struct HorizontalSizeClassEnvironmentKey: EnvironmentKey {
-    static let defaultValue: UserInterfaceSizeClass = .regular
+  public struct HorizontalSizeClassEnvironmentKey: EnvironmentKey {
+    public static let defaultValue: UserInterfaceSizeClass = .regular
   }
 
-  struct VerticalSizeClassEnvironmentKey: EnvironmentKey {
-    static let defaultValue: UserInterfaceSizeClass = .regular
+  public struct VerticalSizeClassEnvironmentKey: EnvironmentKey {
+    public static let defaultValue: UserInterfaceSizeClass = .regular
   }
 
-  extension EnvironmentValues {
+  public extension EnvironmentValues {
     var horizontalSizeClass: UserInterfaceSizeClass {
       get { self[HorizontalSizeClassEnvironmentKey.self] }
       set { self[HorizontalSizeClassEnvironmentKey.self] = newValue }
